Move src, test, docs
