+++ docs/src/LieTheory/cartan_matrix.md
