+++ docs/src/LieTheory/root_systems.md
