+++ docs/src/LieTheory/weight_lattices.md
